Fix SignalResponder signature response when deleted
The test_FnGetAtt_delete_not_found test previously did nothing
because the rsrc.delete() was just getting a generator function
and not actually doing the delete.  This uncovered that we'd throw
an error trying to get the AlarmUrl after resource deletion, so
ensure we return None when the credentials have been deleted.

Change-Id: I7edbaeae26a53e8e424195d2b389b7d45b414da5