Allow choice in how SignalResponder user id is stored.
Some subclasses of SignalResponder may assign a resource_id
with the real reasource that it creates.

With this change, the user_id is always stored in resource data.
If user_id resource data is missing an assumption is made that
this resource was created before this commit, so the
resource_id is used as the user_id.

Subclasses of SignalResponder which previously used the
user ID as their resource_id now need to call resource_id_set
themselves during handle_create.

Change-Id: Iac41ab99f8e9b0953b2cd447ff0e027443fabeca