Correct use of Fn::Split in generated provider templates
The argument to Fn::Split is a list containing the delimiter to split on
and the string to split. The former was omitted when using it in generated
templates.

This is good news, because it looks like nobody is using this feature.
CommaDelimtedLists don't really cut it for list parameters anyway, since
each list item may be e.g. a Map, so in the medium term we should convert
these over to using the Json parameter type in some way.

Change-Id: I02e0833d7032baa092bca97bd015659da07a9bae