Move param format for template-validate to API
This patch is second in a series of patches to unify the code for
property and parameter validation based on a common Schema class.

The current patch moves the formating of parameters during the
template-validate API into a new function in the API layer.
Previously, only a dictionary of internal parameter schema
objects was returned, which meant a close coupling between
the engine's internal data structures and the API expected
format.
With the new format function in the API, we are able to better
separate internal data from API formats and can maintain API
compatibility when refactoring inside the engine.
In a subsequent patch, we can thus move to a consolidated code
base for parameter schema validation (based on the common Schema
class) and don't have to distinguish between CFN and HOT inside
the engine.

Change-Id: I302e58906a0b530d967cbae808d0cf6b2fdcb3bc
Partial-Bug: #1230229
Implements: blueprint schema-code-consolidation