Refactor Parameters Schema based on common Schema
This patch is last in a series of patches to unify the code for
property and parameter validation based on a common Schema class.

The current patch refactors the implementation of the Parameters
Schema to be a sub-class of the common Schema class in constraints.py
so that most of the schema validation code can be shared for property
and parameter validation and no two different code paths for mostly
indentical functionality exist.

Change-Id: I4a1dc2e65e7b118347933bdf18aba49c058aa28b
Closes-Bug: #1230229
Implements: blueprint schema-code-consolidation