Fixed OS::Neutron::Pool creation
Resource OS::Neutron::Pool could not be created when
cookie_name param for session_persistence is not set.

By default params that is not set filled with None
as value and passed to the client. This is incorrent
behaviour in Neutron case. Neutron require that all
fields that is not set should not be passed at all.

So we need to fillter out cookie_name param if it's
value set to None.

Closes-bug: #1270750
Change-Id: Iafb082f69bf1dd7a4f49151f65d5a1b8c69f1337