Native Pseudo Parameters
This is the second patch in order to implement native pseudo parameters
for HOT templates.

The current pseudo-parameters (AWS:: ones) are now mapped to "native"
parameter names.

Now HOTemplate uses HOTParameters instead of Parameters in order to
set up correctly the stack id, which for the HOTemplates is an uuid and
for CFN templates is an arn.

Change-Id: I6361ef1aae55eee762c3f21c9c01ae6ca4a9ff01
Implements: blueprint native-pseudo-params
Closes-Bug: #1229415