Add engine preview stack with Template and Params
Add support for the engine to return the preview of a stack given a
template and the necessary parameters.  No resource will actually be
created, just returned for visualization purposes.  This allows a user
to know beforehand what will happen to their stacks once they used such
template.

The response is a formatted stack plus formatted resources in it:

  {
      'stack': {
        // stack information similar to a call to stack_create/show
        'resources': [
            // resources similar to a call to resource_show
        ]
      }
  }

For now, grouped resources (any of the StackResource) will just be
returned as a single resource, but the preview will be overridden in the
future to allow for more specific behavior.

Implements: blueprint preview-stack (partial)
Change-Id: I8b862110f00ac0cf31d96a86cbc134df895cc19b