heat_keystoneclient add support for stack domain projects
Adds support for creating projects inside the heat stack domain.

A new project will be created in the stack domain if the stack creates
any InstanceUser resource, and the "stack domain project" will then
encapsulate all users associated with resources in the stack.

Adds a new stack_user_domain option which will be used to specify the
name of the domain where heat will create the projects/users.

Change-Id: I7bb1830489baf5fb6761d7e1c914dc3ab0450181
blueprint: instance-users