Map the NotFound exception to HTTPNotFound
Now if the DB API raises a NotFound which propagates
to the REST response, the user will receive a 404.

This seems to be a better default behaviour than a 500,
and other behaviours can be handled by catching NotFound
in service.py

This change is needed for the SoftwareDeployment resource
to do our usual ignore-404-on-handle_delete behaviour.

Change-Id: If00c4de41dc24f462af0b32b29cd6d70f314f3bf