Add handle_update to VolumeAttachment
Updates for volume attachments were handled in UpdateReplace manner,
which fails when in need to attach new volume to the same mount
(or attach the same volume to different mount on the same server,
or attaching the same volume to different server).

This patch makes all properties of VolumeAttachment
to be update_allowed, adds necessary handle_update logic
and fixes VolumeDetachTask to reliably detach old volume
so that new one can be immediately attached.

Closes-Bug: #1269686
Closes-Bug: #1251439
Change-Id: I0aee64036bf8125f76e7f427716412a326ff5eef