Refactor _create_template to not rely on instances
The core of the _create_template logic in autoscaling.py is now in
a separate static method, _resource_templates. _resource_templates is
destined to be moved to another scaling module that is intended to
be used outside of the heat engine, so it won't have access to instance
objects.

Implements: partial blueprint as-lib

Change-Id: I0eda024c88534207daa71de2a8eb53738cdd89ef