Do not log error on software config/deployment NotFound
This change decorates software config methods so that errors
are not logged if a NotFound is raised on a software config/deployment
operation.

An error should not be logged as any user can fetch a resource
that does not exist. There are most likely other service methods
which require this decorator too.

Change-Id: I39cd52925006bddccf55ca1a9f43e0ab37e27b14