migrate User/AccessKey resources to StackUser base class
migrate the User/AccessKey resources to use the StackUser
base class, and therefore convert to stack domain users.

Note that becuse the User resource inherits from StackUser,
and the keypair is managed by that resource now, we have
to update both resources at the same time, as it's not
possible to update individually and keep them both working.

This (finally! :) resolves bug #1089261 and allows all
resources to be created even if the user creating the
stack doesn't have the required roles to create keystone
users.

Closes-Bug: #1089261
blueprint: instance-users

Change-Id: I854644e7669150521f8904cdb76c292519ef4167