RPC method to fetch deployments metadata
This method returns a list of software configs which are currently
deployed to a given server. Each config has the deployment input values
and current action inserted into it. This means that each config is
ephemeral and derived from the canonical config specified in the deployment
resource. The list of configs is sorted by config name. This allows the
template to control what order the configs are returned by setting the
'name' attribute of each deployment resource.

Having a separate RPC (and REST) method for deployment metadata also means
that servers will eventually be able to poll for deployment metadata directly,
avoiding the need to parse the stack for every poll.

partial blueprint hot-software-config-rest

Change-Id: I374f1e15d5255a3ad1264859fd845265cc60f218