REST deployment metadata method
A GET on the path /software_deployments/metadata/<server_id>
will return a list of software configs which should be deployed
to that server.

See the previous commit for the justification for this method.

In the future servers will be able to poll for this directly so
the policy.json does not have the deny_stack_user rule. Now that
this method exists software_deployments:index can now enforce
deny_stack_user.

partial blueprint hot-software-config-rest

Change-Id: I514677850cfb742bec85a5db75c71e6b2c7c9ea2