Change Resource timestamps to save correct info
Resource.updated_time is currently being treated as modification time
of the object in the database.  Since it's not a very useful thing to
have and the CFN API treats this value as "the last time an update
call was issued", it makes sense to have the same behavior here as
well.

This changes the current real-time retrieval of created_time and
updated_time in favor of the attributes already present in the
resource and changes the information stored in these two fields to be
the time the resource was created and the time the resource was last
updated by the user, respectively.

Co-Authored-By: Richard Lee <rblee88@gmail.com>
Related-Bug: #1193269
Change-Id: Iad689b97d237fa63856b878827a1fcb1676c991c