Server property for software config transport
Setting a value for software_config_transport will change the configured
transport which is used to get software config metadata to the server.

The default value POLL_SERVER_CFN allows the current behaviour of using
an EC2 keypair signed request to poll for metadata changes.

The new value POLL_SERVER_HEAT will allow heatclient (or equivalent) to
fetch metadata for the server resource via a standard native API resource-show
request.

os-collect-config is configured with the (currently fictional) heat_server_poll
collector and provided with credentials which can be used to get a valid auth token.

The next most likely transport is one which polls the deployments metadata API
directly. Hopefully in the future there will be non-polling based transports.

Change-Id: Ib6fd76ef58e2b8e11b62d127e1df654254cd15ec