Use stack ID to create domain project name
If the project for a given stack didn't get deleted for any reason then
launching a stack with the same name will fail as project names
are enforced to be unique.

By using the stack ID instead this issue is avoided if the project delete
didn't occur.

The resulting project name is truncated to the limit of 64.

Change-Id: Iba09cf94bdaa6a92d9520ed1743805862dc1b14d
Closes-Bug: #1288514