Refactor stack watching into a separate class
In later oslo.messaging patches, the Stack watch functionality must
be called from two separate places (the endpoint class and the service
class).  In order to do that and not result in needless code
duplication, break it into a separate class and so the object can
be passed around.

Change-Id: I4f4fae0182779d955ac0e15109fc641b271ddc01
blueprint: oslo-messaging