Never specify project_name in keystone v3 auth
All of our middlewares populate the context with both the tenant name and the
tenant_id. The tenant_id, is unique across domains which means you can do lookups
without having to know what domain the project is in.

When project_id and project_name are both specified, it seems that project_name
takes precedence, which will fail when the credentials are from the stack domain since
we cannot lookup by project names when the domain is not specified.

It looks like specifying project_name is never the right thing to do, and we should
assume that the context always has a tenant_id.

Change-Id: Ifea21e13be482a6347438d6baaca3111caad4a97
Closes-Bug: #1290621