Fail if non-existent security group referenced
When using AWS::EC2::SecurityGroup resources with nova-network (not
Neutron), if a SourceSecurityGroupName was specified in a rule but the
name did not correspond to an extant security group it would be silently
ignored. This change ensures that any such error in the template will cause
creation of the security group to fail.

Closes-bug: #1291091

Change-Id: I9344451c83f4184ef46d4e9906b94abda08836df