Don't re-bind parameters during stack update
While we're in the process of updating a stack, we set the stack's parameters
to the new, updated values. However, we don't want to change existing
resources' idea of their own values until we have explicitly done an update of
them to bring them into line with the new template/parameters. So when parsing
a template snippet, store the parameters object to which it will refer.
References will resolve to the new parameters only when the template has been
re-resolved.

Closes-bug: #1291411

Change-Id: Iebbb518c5d15cfba858db001648f3412a7d8b1c0