Migrate invalid template version values
If a stack was created with a template that specified an invalid time,
a variety of operations cannot be done on the stack, resulting in a
general breakage of Heat.  This patch migrates any occurances of version
numbers that are invalid that were specified prior to the merging of this
patch.  This patch also handles invalid version fields where a non-date
is specified or an invalid date was specified.

Change-Id: I5387ffb84dba5160103b3cb02f3d1cc5f352ff29
Closes-bug: 1296859